updated filter parameters, added to PermStack
Filter function parameters have been made constant to indicate that they do not modify or keep any vectors sent to them. PermStack had a RoundsCurrentlyBeingTested variable added to hold which rounds are currently being tested so that when rounds are popped off the stack the numbers aren't lost. A deconstructor was added that just empties the stack and deletes the round vectors. clear stack method created.