added to getValidOrder
added more logic for when the algorithm is at the deepest depth but does not satisfy the part of the problem that requires books only be sent to each person once. logic for determinging that each person recieved each book once not yet implimented. current logic for if the solution is suboptimal is not yet finished.