updated readme with research
results from research lead me to latin squares generation (which was a dead end due to the fact that they are just as expensive to generate). still want to look into euler/hamiltonian path but having each node be who sent to who for one specific person didnt result in a graph that explained what was shown in testing. number of states is too many to try doing euler/hamiltonian on the entire thing using pen and paper for any graph with p>3. also if this simplifies to a hamiltonian path it wouldnt even help as hamiltonian path is np-hard, finding if each # of players would take just as long as testing every possible solution.